# auth-service/src/main/resources/openapi/auth-api.yml
openapi: 3.1.0
servers:
  - description: Service for Auth API
    url: https://me.com/auth/1.0.0
info:
  description: This is an Auth API
  version: "1.0.1"
  title: Auth API
  contact:
    email: igmartinez87@gmail.com
tags:
  - name: auth
    description: auth services

paths:

  /auth/login:
    post:
      tags:
        - auth
      summary: user login
      operationId: userLogin
      description: endpoint to login and get a token
      responses:
        '200':
          description: user logged and token created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenReponse'
        '400':
          description: bad input parameter
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
        description: User data to login

  /auth/refresh:
    post:
      tags:
        - auth
      summary: token refresh
      operationId: tokenRefresh
      description: endpoint to refresh access
      responses:
        '200':
          description: both tokens refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenReponse'
        '400':
          description: bad input parameter
      requestBody:
        content:
          application/json:
            schema:
              type: string
              examples: [eyASIOJQEGeGOJeqgjIEqawjqweIGqWEGJ]
        description: refresh token

    #/auth/logout:
    #post:

  /users:
    post:
      tags:
        - auth
      summary: registers user
      operationId: registerUser
      description: endpoint to register an user
      responses:
        '201':
          description: user created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: bad input parameter
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
        description: User to register


components:
  schemas:

    UserRequest:
      type: object
      required:
        - username
        - password
        - mail
        - birthdate
        - is_active
      properties:
        username:
          type: string
          examples: [Pedro]
          min: 8
        password:
          type: string
          pattern: ^(?=.*\d)(?=.*[a-z])(?=.*[A-Z]).{10,}$
          examples: [Pedro]
        mail:
          type: string
          format: email
          examples: [mail@mail.com]
        birth_date:
          type: string
          examples: [12/12/2012]
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleRequest'
        is_active:
          type: boolean
          examples: [true]

    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          examples: [1]
        username:
          type: string
          examples: [Pedro]
        password:
          type: string
          examples: [theworstpasswordintheworld]
        mail:
          type: string
          format: email
          examples: [mail@mail.com]
        birth_date:
          type: string
          examples: [12/12/2012]
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleRequest'
        is_active:
          type: boolean
          examples: [true]

    RoleRequest:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
          examples: [1]
        name:
          type: integer
          format: int64
          examples: [ROLE_ADMIN, ROLE_USER]

    RoleResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          examples: [1]
        name:
          type: integer
          format: int64
          examples: [ROLE_ADMIN, ROLE_USER]

    TokenReponse:
      type: object
      properties:
        token:
          type: string
          examples: [eyASIOJQEGeGOJeqgjIEqawjqweIGqWEGJ]
        refresh_token:
          type: string
          examples: [eyASIOJQEGeGOJeqgjIEqawjqweIGqWEGJ]